+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJSC9S2BSJFMQHTX5VGD6R9N"
+        "01KJSJH629GMSQBF89M3XCMWKT"
     )
     r"""The unique identifier of the data source"""