+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJSC9S2E3G9GEVXDGTM9DEW5"
+        "01KJSJH62BZ34RDCZ7J1YJ009B"
     )
     r"""The unique identifier of the data source"""