+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJSC9S2KDY8ET1KX2X7RVGFY"
+        "01KJSJH62HQQ818V6MNVB27HF1"
     )
     r"""The unique identifier of the data source"""