+++ packages/orq-rc/RELEASES.md
@@ -6269,3 +6269,13 @@ Based on:
 - [python v4.5.0-rc.22] packages/orq-rc
 ### Releases
 - [PyPI v4.5.0-rc.22] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.22 - packages/orq-rc
+
+## 2026-03-03 10:07:53
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.739.0 (2.846.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.5.0-rc.23] packages/orq-rc
+### Releases
+- [PyPI v4.5.0-rc.23] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.23 - packages/orq-rc